modprobe.d: Set RAID0 default layout
The RAID0 layout changed from pre- to post-Linux-3.14
without any versioning information, which was only added
recently in mdadm. Therefore, when a RAID0 setup lacks the
versioning information we need to define what the default is,
or otherwise the kernel refuses to operate it.
Chose the alternate layout because we care about setups that
are created with post-3.14 kernels which are in use for the
last five years already.