scripts: Allow overriding the base
When the second parameter is not provided, the script works as it used
to.

The reason for the change is to be able to run the script before
anything is provisioned in the new root filesystem. This will populate
the /etc/passwd and /etc/group files therein, so systemd-tmpfiles can
resolve the user and group information used in tmpfiles. This is a
change in systemd made in 246, as it only resolves the user and group
information from databases inside the overridden root filesystem (with
the --root flag).

This also extends the groups with the root and core groups to be
copied over, because these groups are needed by systemd-tmpfiles too.