tmpfiles: Ensure that /etc/coreos to flatcar symlink exists
The entries from the update or locksmith directive of a Container Linux
Config currently use the /etc/coreos/ folder for Ignition files:
  printf "locksmith:\n  reboot_strategy: off" | ct
results in:
  {"ignition":{"config":{},"security":{"tls":{}},"timeouts":{},
  "version":"2.3.0"},"networkd":{},"passwd":{},"storage":
  {"files":[{"filesystem":"root", "path":"/etc/coreos/update.conf",
  "contents":{"source":"data:,%0AREBOOT_STRATEGY%3D%22off%22",
  "verification":{}},"mode":420}]},"systemd":{}}
With the pre-built rootfs this works because /etc/coreos is a symlink
to /etc/flatcar. However, with a fresh rootfs created by Ignition this
results in a /etc/coreos folder created by Ignition because there is no
symlink and the resulting files are ignored since only /etc/flatcar/ is
considered by update-engine and locksmith.
Set up the symlink and the target folder through a tmpfile directive to
ensure that it is there when Ignition runs even when with a new rootfs.

Fixes https://github.com/flatcar-linux/Flatcar/issues/190