tmpfiles.d/baselayout.conf: Add rules for /opt and /opt/bin
The well-known path /opt/bin is used in PATH as place for extra
binaries and /opt in general is a place for additional software.
Containerd creates /opt on runtime if it is not present but uses
restrictive permissions, causing non-root programs to fail accessing,
e.g., /opt/bin.
Add tmpfile directives for /opt and /opt/bin to ensure that they
have the correct permissions.

Fixes https://github.com/kinvolk/Flatcar/issues/279