sysctl.d: Add prefix 60 to baselayout.conf and set rp_filter values
The baselayout.conf file was applied after all "[0-9]+.*.conf" files
due to sorting by file name. This made it difficult for users to apply
their configuration files because they had to use a name starting with
something being sorted after "bb" while everybody expects "90" to be
the latest in the order. This also caused that baselayout was
overwriting the "50-default.conf" file from systemd.
Sort the baselayout file after the "50-default.conf" file but before
any possible user files like "90-...". By inserting the baselayout file
after "50-default.conf" we have a good way to overwrite the systemd
settings for rp_filter that we want to change to 0. This allows us to
drop any patches to the systemd ebuild file.