passwd: use correct GID for tss
The tss GID is not 236 but 252. Luckily the tss group explicitly put
the user into it but still we don't want an unallocated group to be
used. While for consistency it would be nice to choose 236 as new GID,
this could be incompatible with any files created under /var/ or a
similar case where the old GID is still on the filesystem. With the
correct GID referenced, the explicit add of the user inside the group
file is not needed anymore.