profile: filter out bullet point when parsing failed units
The output of "systemctl list-units --state=failed --no-legend" still
contains the bullet point which is not expected and ends up being taken
as the unit name which was previously on the start of the line.
Filter the bullet point out to stay compatible with the old behavior in
case upstream would remove the bullet point again.