baselayout/passwd|group: remove duplicates of acct-user|group
With the new mechanism introduced in
https://github.com/flatcar-linux/scripts/pull/227
we don't need to duplicate the Gentoo acct-user|group ID allocations
in our database anymore because the systemd-sysuser accounts now get
appended to it on build time.
Unfortunately we diverge with some IDs from Gentoo's acct-user and
acct-group allocations and can not remove more user entries but have to
keep them around to ensure that the user IDs are stable for updating
users.