*: Reorganize the files
Sort the files and directories that the Makefile installs into
categories, so that everything that gets installed to `/etc` is inside
`etc` directory. Same for `/usr/bin` and `bin`, `/usr/lib` and `lib`,
and `/usr/share` and `share`.

The files in `scripts` (one script and one systemd unit) were moved to
`bin` and to `share`. For SDK builds these will be removed before
installation, which is why the `bin` part is made optional in
Makefile.