dumb-tmpfiles-proc.sh: Add a dumb tmpfiles processor script
It's a temporary solution, which will be dropped once we migrate to
Gentoo's baselayout. It is quite a dubm and simple processor, because
it is very limited in its functionality compared to systemd's tmpfiles
processor, and it requires the user and group to be passed up front
through the environment variables. This is because the script is
tailored to be used only in the ebuild, when installing baselayout in
src_install and pkg_preinst phases.

In src_install the script will be used to install the files and
symlinks, together with their parent directories, thus excluding the
empty directories. These will get created in the pkg_preinst phase.

The point of adding the script was to stop using systemd's tmpfiles
processor, which requires user and group information to exist in
/etc/passwd or /etc/group. Obviously it would be even better to not
abuse the tmpfiles for the package installation purposes, but I think
that this will come with the switch to Gentoo's baselayout.